+++ integrations/gitlab-v2/clients/helpers/utils/exceptions.py
@@ -1,5 +1,2 @@
-from typing import Any
-
-
 class GraphQLQueryError(Exception):
     pass
